first pass at a doc around code and design
